Avoid stashing source if we're not sure it's source; relocate tokbuf and strbuf
The previous method of keeping EasyAsm in bank 5 allowed the user to indicate whether program memory contains source code based on the menu option they choose. This new method of always stashing program memory to install EasyAsm for every invocation makes it easier to accidentally overwrite stashed source with binary data. This is now tracked with a dirty flag.

The dirty flag comes with a caveat, which I added to the README for now: if the user assembles and tests a program, exits uncleanly, then does something other than re-invoke EasyAsm to restore source, EasyAsm will assume program memory is dirty upon next invocation, even if the user "cleaned" it manually.

In the future, it might make more sense for EasyAsm to have multiple program modules invoked from a menu utility that runs from bank 5 as before. It can then support the previous workflow, while still taking advantage of bank 0 program memory. (Program modules might never be necessary otherwise, but imagine an EasySuite of tools that can be invoked from the menu that might outgrow 44 KB.)